Stop pamd spamming the journal in Archlinux.
Recent updates in pam (pam v1.4+) introduced the
pam_systemd_homed.so module. The way system-auth is structured
is causing pam_systemd_homed.so to spam the journal whenever
sudo is used.

With these twekas, systemd_homed.so is completely skipped since
I don't use it. Since the systemd_homed.so line is getting
commented out, pam_unix.so's control must also be changed so
it only skips one line and jumps to pam_permit.so.