Workaround tags and gather facts.
At the moment, if one of the tags that applies to
gather-facts.yml is skipped (e.g., --skip-tags=gnome),
the tasks inside gather-facts.yml will never run, causing
conditionals to fail all over the place.

To fix this, I've included gather-facts.yml in every play that
requires to know what packages are installed.