Refactor how the GNOME setup is done.
As of 2cd6559, gather-facts.yml is no longer included in the main
playbook, but in every play itself, meaning the conditional

"'gnome-shell' in installed_packages
or 'gnome-shell-performance' in installed_packages"

in 'include_tasks: gnome-setup.yml' will fail.

To avoid that, I've removed the conditional from main.yml and
wrapped the entire gnome configuration in a `block`, that
uses this same conditional.